# swagger/swagger.json
{
   "openapi": "3.0.0",
   "info": {
     "title": "CAB230: Logan Allen - N10138790",
     "description": "Crime statistics and data based upon the assessment for CAB230 - Sem1 - 2019.",
     "version": "1.0"
   },
   "servers": [
     {
       "url": "https://localhost:80"
     }
   ],
   "paths": {
     "/register": {
       "post": {
         "tags": [
           "Authentication"
         ],
         "summary": "Registers a new user account",
         "description": "This is your starting point.",
         "requestBody": {
           "required": true,
           "content": {
             "application/json": {
               "schema": {
                 "type": "object",
                 "properties": {
                   "email": {
                     "type": "string",
                     "required": true,
                     "example": "user@gmail.com"
                   },
                   "password": {
                     "type": "string",
                     "required": true,
                     "example": "userpassword"
                   }
                 }
               }
             }
           }
         },
         "responses": {
           "201": {
             "description": "Successful",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "message": {
                       "type": "string",
                       "example": "Account successfully registered."
                     }
                   }
                 }
               }
             }
           },
           "400": {
             "description": "Unsuccessful",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "message": {
                       "type": "string",
                       "example": "The information you supplied already exists."
                     }
                   }
                 }
               }
             }
           }
         }
       }
     },
     "/login": {
       "post": {
         "summary": "Login with an existing user account",
         "requestBody": {
           "required": true,
           "content": {
             "application/json": {
               "schema": {
                 "type": "object",
                 "properties": {
                   "email": {
                     "type": "string",
                     "required": true,
                     "example": "user@gmail.com"
                   },
                   "password": {
                     "type": "string",
                     "required": true,
                     "example": "userpassword"
                   }
                 }
               }
             }
           }
         },
         "responses": {
           "200": {
             "description": "OK",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "access_token": {
                       "type": "string",
                       "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjo1Nzg2LCJlbWFpbCI6ImRlbW8xMjM0QGdtYWlsLmNvbSJ9LCJpYXQiOjE1NTUzMjQ2MDIsImV4cCI6MTU1NTQxMTAwMn0.mo85iKuAQ_h8n8-6pkeNNs5CCnV0G8hScxhi_ml_Q6c"
                     },
                     "token_type": {
                       "type": "string",
                       "example": "Bearer"
                     },
                     "expires_in": {
                       "type": "intager",
                       "example": 86400
                     }
                   }
                 }
               }
             }
           },
           "401": {
             "description": "Unauthorized",
             "content": {
               "application/json": {
                 "schema": {
                   "type": "object",
                   "properties": {
                     "message": {
                       "type": "string",
                       "example": "Invalid login."
                     }
                   }
                 }
               }
             }
           }
         },
         "tags": [
           "Authentication"
         ]
       }
     },
...
   },
...
 }